typescript refinements
