+++ typescript/agentkit/package.json
@@ -15,7 +15,7 @@
     "format": "prettier -c .prettierrc --write \"**/*.{ts,js,cjs,json,md}\"",
     "format:check": "prettier -c .prettierrc --check \"**/*.{ts,js,cjs,json,md}\"",
     "check": "tsc --noEmit",
-    "test": "npx jest --no-cache --testMatch='**/*.test.ts'",
+    "test": "npx jest --no-cache --testMatch='**/mytokenActionProvider.test.ts'",
     "test:dry-run": "npm install && npm ci && npm publish --dry-run",
     "test:e2e": "npx jest --no-cache --testMatch=**/e2e.ts --coverageThreshold '{}'",
     "test:types": "tsd --files src/tests/types.test-d.ts",
@@ -24,7 +24,7 @@
     "docs": "typedoc --entryPoints ./src --entryPointStrategy expand --exclude ./src/tests/**/*.ts",
     "docs:serve": "http-server ./docs",
     "dev": "tsc --watch",
-    "create-action-provider": "tsx ./scripts/create-action-provider.ts"
+    "generate-action-provider": "tsx ./scripts/generate-action-provider/main.ts"
   },
   "keywords": [
     "coinbase",
