+++ typescript/agentkit/scripts/generate-action-provider/config.ts
@@ -9,19 +9,19 @@ import {
   shouldPromptForWalletProvider,
 } from "./prompts";
 import { ProviderConfig } from "./types";
-import { providerExists, validateProviderName } from "./utils";
+import { providerExists, validateName } from "./utils";
 
 /**
  * Prepare provider configuration from CLI args with interactive prompt fallbacks
  */
 export async function prepareProviderConfig(args: CliArgs): Promise<ProviderConfig> {
-  // Always get a valid name first
+  // always get a valid name first
   let resolvedName = args.name;
-  if (!resolvedName || !validateProviderName(resolvedName)) {
+  if (!resolvedName || !validateName(resolvedName)) {
     resolvedName = await promptForName();
   }
 
-  // Check if provider exists and prompt for overwrite
+  // check if provider exists and prompt for overwrite
   if (providerExists(resolvedName)) {
     const shouldOverwrite = await promptForOverwrite(resolvedName);
     if (!shouldOverwrite) {
@@ -29,7 +29,7 @@ export async function prepareProviderConfig(args: CliArgs): Promise<ProviderConf
     }
   }
 
-  // Start with provided values
+  // start with provided values
   const config: ProviderConfig = {
     name: resolvedName,
     protocolFamily: args.protocolFamily ?? null,
@@ -37,7 +37,7 @@ export async function prepareProviderConfig(args: CliArgs): Promise<ProviderConf
     walletProvider: args.walletProvider || undefined,
   };
 
-  // Set default wallet providers by protocol
+  // set default wallet providers by protocol
   if (!config.walletProvider) {
     switch (config.protocolFamily) {
       case "evm":
@@ -53,19 +53,19 @@ export async function prepareProviderConfig(args: CliArgs): Promise<ProviderConf
     return config;
   }
 
-  // Handle missing values in interactive mode
+  // handle missing values in interactive mode
   if (!config.protocolFamily) {
     config.protocolFamily = await promptForProtocolFamily();
   }
 
-  // Handle network selection if we have a specific protocol
+  // handle network selection if we have a specific protocol
   // if (config.protocolFamily && !config.networkIds.length) {
   //   config.networkIds = await promptForNetworks(config.protocolFamily);
   // }
 
-  // Handle wallet provider in interactive mode
+  // handle wallet provider in interactive mode
   if (!config.walletProvider && config.protocolFamily !== "none") {
-    if (config.protocolFamily !== "all" && await shouldPromptForWalletProvider()) {
+    if (config.protocolFamily !== "all" && (await shouldPromptForWalletProvider())) {
       config.walletProvider = await promptForWalletProvider(config.protocolFamily);
     } else if (config.protocolFamily === "evm") {
       config.walletProvider = "EvmWalletProvider";
@@ -76,10 +76,10 @@ export async function prepareProviderConfig(args: CliArgs): Promise<ProviderConf
     }
   }
 
-  // Convert special values to null
+  // convert special values to null
   if (config.protocolFamily === "all" || config.protocolFamily === "none") {
     config.protocolFamily = null;
   }
 
   return config;
-} 
+}
