+++ typescript/agentkit/scripts/generate-action-provider/templates/actionProvider.test.ts.template
@@ -49,7 +49,7 @@ describe("{{name_pascal}}ActionProvider", () => {
   });
 
   describe("network support", () => {
-    {% if protocol_family == "all" %}
+    {% if not protocol_family %}
     it("should support all network types", () => {
       expect(
         provider.supportsNetwork({
@@ -60,7 +60,7 @@ describe("{{name_pascal}}ActionProvider", () => {
       ).toBe(true);
     });
     {% endif %}
-    {% if protocol_family != "all" and not networkIds|length %}
+    {% if protocol_family and not networkIds|length %}
     it("should support the protocol family", () => {
       expect(
         provider.supportsNetwork({
@@ -77,7 +77,7 @@ describe("{{name_pascal}}ActionProvider", () => {
       ).toBe(false);
     });
     {% endif %}
-    {% if protocol_family != "all" and networkIds|length %}
+    {% if protocol_family and networkIds|length %}
     it("should support configured networks", () => {
       {% for networkId in networkIds %}
       {% if protocol_family == "evm" %}
@@ -115,7 +115,7 @@ describe("{{name_pascal}}ActionProvider", () => {
       ).toBe(false);
     });
     {% endif %}
-    {% if protocol_family != "all" %}
+    {% if protocol_family %}
 
     it("should handle invalid network objects", () => {
       expect(provider.supportsNetwork({} as Network)).toBe(false);
@@ -135,7 +135,6 @@ describe("{{name_pascal}}ActionProvider", () => {
       const validInput = {
         fieldName: "test",
         amount: "1.0",
-        enumField: "option1" as const,
       };
       const parseResult = ExampleActionSchema.safeParse(validInput);
       expect(parseResult.success).toBe(true);
@@ -145,7 +144,6 @@ describe("{{name_pascal}}ActionProvider", () => {
       const invalidInput = {
         fieldName: "",
         amount: "invalid",
-        enumField: "invalid",
       };
       const parseResult = ExampleActionSchema.safeParse(invalidInput);
       expect(parseResult.success).toBe(false);
@@ -158,7 +156,6 @@ describe("{{name_pascal}}ActionProvider", () => {
       const args = {
         fieldName: "test",
         amount: "1.0",
-        enumField: "option1" as const,
       };
       const result = await provider.exampleAction(mockWalletProvider, args);
       expect(result).toContain(args.fieldName);
@@ -171,7 +168,6 @@ describe("{{name_pascal}}ActionProvider", () => {
       const args = {
         fieldName: "test",
         amount: "1.0",
-        enumField: "option1" as const,
       };
       const result = await provider.exampleAction(args);
       expect(result).toContain(args.fieldName);
