+++ typescript/agentkit/scripts/generate-action-provider/templates/index.ts.template
