+++ typescript/agentkit/scripts/generate-action-provider/types.ts
@@ -4,13 +4,9 @@ import { ProtocolFamily, NetworkId, WalletProvider } from "./constants";
  * Configuration for an action provider
  */
 export interface ProviderConfig {
-  /** The name of the provider */
   name: string;
-  /** The protocol family (e.g., "evm", "svm") or null for all protocols */
   protocolFamily: ProtocolFamily | null;
-  /** The network IDs (e.g., ["ethereum-mainnet", "base-mainnet"]) */
   networkIds: NetworkId[];
-  /** The wallet provider to use (optional) */
   walletProvider?: WalletProvider;
 }
