--- typescript/agentkit/scripts/templates/schemas.ts.template
@@ -1,45 +1,0 @@
-/**
- * Schemas for {{name}} Action Provider
- *
- * This file contains the Zod schemas that define the input types
- * for the {{name}} action provider's actions.
- *
- * @module {{name}}/schemas
- */
-
-import { z } from "zod";
-
-/**
- * Action schemas for the {{name}} action provider.
- *
- * This file contains the Zod schemas that define the shape and validation
- * rules for action parameters in the {{name}} action provider.
- */
-
-/**
- * Example action schema demonstrating various field types and validations.
- * Replace or modify this with your actual action schemas.
- */
-export const ExampleActionSchema = z
-  .object({
-    /**
-     * A descriptive name for the field
-     */
-    fieldName: z.string().min(1).max(100),
-
-    /**
-     * The amount to use in the action (as a decimal string, e.g. "0.01")
-     */
-    amount: z.string().regex(/^\d*\.?\d+$/, "Amount must be a valid decimal number"),
-
-    /**
-     * Optional parameter example
-     */
-    optionalField: z.string().optional(),
-
-    /**
-     * Constrained choices example
-     */
-    enumField: z.enum(["option1", "option2", "option3"]),
-  })
-  .describe("Example action parameters");
